react context only in appd.js and Dashboard.js
I really should have committed before starting this, as it's still possible that it will backfire. I'm committing now because I did a fair amount of sleuthing to figure out soil textures that I don't want to lose, and the initial setup of weather. If I have to reset back to the previous commit I'll lose all of that.